tests/requirements.txt
# Core Testing Framework
pytest==7.4.3
pytest-html==4.1.1
pytest-json-report==1.5.0
pytest-xdist==3.5.0  # For parallel test execution

# HTTP Client
requests==2.31.0

# Test Data Generation
faker==22.0.0

# Database Access
pymongo==4.6.1

# Environment Configuration
python-dotenv==1.0.0

# Utilities
colorama==0.4.6  # For colored terminal output
tabulate==0.9.0  # For formatted output tables
